requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
isort>=5.12.0
mypy>=1.5.1
logging==0.4.9.6
